+++ include/solar-system-clock/layers/sun.h
@@ -14,7 +14,7 @@ namespace solarsystemclock::layers {
         ~Sun();
 
         void resize(int width, int height) override;
-        void update(float dt) override;
+        void update(double dt) override;
         void draw() override;
 
     private:
