+++ lib/layers/orbitrings.cpp
@@ -17,13 +17,13 @@ OrbitRingss::~OrbitRingss() {
     delete m_texture;
 }
 
-float OrbitRingss::gap() const {
-    float scale = m_size / 1024.0;
+double OrbitRingss::gap() const {
+    double scale = m_size / 1024.0;
     return 55.f * scale;
 }
 
 int OrbitRingss::radius(int orbit) const {
-    float scale = m_size / 1024.0;
+    double scale = m_size / 1024.0;
     int start = 100 * scale;
 
     return start + gap() * orbit;
